Expose constructor
This allows for efficient serialisation.
