+++ ChangeLog.md
@@ -3,3 +3,7 @@
 ## 0.1
 
 * First version. Released on an unsuspecting world.
+
+## 0.2
+
+* Expose constructor in `Data.Text.Short.Internal` module
