+++ text-short.cabal
@@ -22,6 +22,7 @@ Source-Repository head
 library
   default-language:    Haskell2010
   exposed-modules:     Data.Text.Short
+                       Data.Text.Short.Internal
   build-depends:       base        >= 4.7    && < 4.11
                      , bytestring  >= 0.10.4 && < 0.11
                      , hashable    >= 1.2.6  && < 1.3
